feat: move functionality from App.jsx into ClickMe (own component)
